requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0
wandb>=0.12.0
transformers>=4.20.0
librosa>=0.9.0
opencv-python>=4.5.0
scipy>=1.7.0
pyyaml>=5.4.0
tensorboard>=2.8.0
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
opacus>=1.0.0
h5py>=3.0.0
